+++ packages/instrumentation-elasticsearch/test/elastic.spec.ts
@@ -1,8 +1,7 @@
 import 'mocha';
 import nock from 'nock';
 import { expect } from 'chai';
-import { InMemorySpanExporter, SimpleSpanProcessor } from '@opentelemetry/tracing';
-import { NodeTracerProvider } from '@opentelemetry/node';
+import { getTestSpans } from 'opentelemetry-instrumentation-testing-utils';
 import { ElasticsearchInstrumentation } from '../src/elasticsearch';
 
 const instrumentation = new ElasticsearchInstrumentation();
@@ -13,22 +12,12 @@ const esNock = nock(esMockUrl);
 const client = new Client({ node: esMockUrl });
 
 describe('elasticsearch instrumentation', () => {
-    const provider = new NodeTracerProvider();
-    const memoryExporter = new InMemorySpanExporter();
-    const spanProcessor = new SimpleSpanProcessor(memoryExporter);
-    provider.addSpanProcessor(spanProcessor);
-    instrumentation.setTracerProvider(provider);
-
     before(() => {
         instrumentation.enable();
     });
 
     after(() => {
         instrumentation.disable();
-    });
-
-    beforeEach(() => {
-        memoryExporter.reset();
     });
 
     it('should create valid span', async () => {
@@ -48,7 +37,7 @@ describe('elasticsearch instrumentation', () => {
             index: 'the-simpsons',
         });
 
-        const spans = memoryExporter.getFinishedSpans();
+        const spans = getTestSpans();
         expect(spans?.length).to.equal(2);
         expect(spans[0].attributes).to.deep.equal({
             'db.system': 'elasticsearch',
@@ -75,7 +64,7 @@ describe('elasticsearch instrumentation', () => {
         esNock.get('/_cluster/settings').reply(200, {});
 
         await client.cluster.getSettings();
-        const spans = memoryExporter.getFinishedSpans();
+        const spans = getTestSpans();
 
         expect(spans?.length).to.equal(1);
         expect(spans[0].attributes).to.deep.equal({
@@ -94,7 +83,7 @@ describe('elasticsearch instrumentation', () => {
         instrumentation.disable();
         await client.cluster.getSettings();
         instrumentation.enable();
-        const spans = memoryExporter.getFinishedSpans();
+        const spans = getTestSpans();
         expect(spans?.length).to.equal(0);
     });
 });