+++ packages/instrumentation-express/test/opentelemetry-express-layers.spec.ts
@@ -1,22 +1,15 @@
 import 'mocha';
 import expect from 'expect';
-// initialize the provider before importing the instrumented package
-import { NodeTracerProvider } from '@opentelemetry/node';
-import { context, ContextManager } from '@opentelemetry/api';
-const provider = new NodeTracerProvider();
 
 import { ExpressInstrumentation } from '../src';
 import { AddressInfo } from 'net';
-import { InMemorySpanExporter, SimpleSpanProcessor, ReadableSpan } from '@opentelemetry/tracing';
-import { AsyncHooksContextManager } from '@opentelemetry/context-async-hooks';
+import { ReadableSpan } from '@opentelemetry/tracing';
 
 const instrumentation = new ExpressInstrumentation();
-instrumentation.setTracerProvider(provider);
 
 import axios from 'axios';
 import express from 'express';
 import * as http from 'http';
-import { ExpressInstrumentationAttributes } from '../src/types';
 import {
     errorMiddleware,
     expectRouteAttributes,
@@ -30,11 +23,7 @@ import {
 } from './utils';
 import { describe } from 'mocha';
 
-describe('aspecto-opentelemetry-express-layers', () => {
-    const memoryExporter = new InMemorySpanExporter();
-    const spanProcessor = new SimpleSpanProcessor(memoryExporter);
-    provider.addSpanProcessor(spanProcessor);
-    let contextManager: AsyncHooksContextManager;
+describe('opentelemetry-express-layers', () => {
     let app: express.Application;
 
     const sendRequest = async (urlPath: string): Promise<ReadableSpan> => {
@@ -52,7 +41,7 @@ describe('aspecto-opentelemetry-express-layers', () => {
                     }
 
                     server.close();
-                    resolve(getExpressSpans(memoryExporter)[0]);
+                    resolve(getExpressSpans()[0]);
                 } catch (err) {
                     console.log(err);
                 }
@@ -61,17 +50,11 @@ describe('aspecto-opentelemetry-express-layers', () => {
     };
 
     before(() => {
-        contextManager = new AsyncHooksContextManager();
-        context.setGlobalContextManager((contextManager.enable() as unknown) as ContextManager);
         instrumentation.enable();
     });
 
     beforeEach(() => {
         app = express();
-    });
-
-    afterEach(() => {
-        memoryExporter.reset();
     });
 
     it('no routes registered', async () => {