+++ packages/instrumentation-kafkajs/test/kafkajs.spec.ts
@@ -1,11 +1,10 @@
 import 'mocha';
-import { KafkaJsInstrumentation, KafkaJsInstrumentationConfig } from '../src';
-import { InMemorySpanExporter, SimpleSpanProcessor, ReadableSpan } from '@opentelemetry/tracing';
-import { NodeTracerProvider } from '@opentelemetry/node';
-import { AsyncHooksContextManager } from '@opentelemetry/context-async-hooks';
-import { context, propagation, SpanKind, SpanStatusCode, Span, ContextManager } from '@opentelemetry/api';
-import { MessagingDestinationKindValues, SemanticAttributes } from '@opentelemetry/semantic-conventions';
 import expect from 'expect';
+import { KafkaJsInstrumentation, KafkaJsInstrumentationConfig } from '../src';
+import { ReadableSpan } from '@opentelemetry/tracing';
+import { propagation, SpanKind, SpanStatusCode, Span } from '@opentelemetry/api';
+import { MessagingDestinationKindValues, SemanticAttributes } from '@opentelemetry/semantic-conventions';
+import { getTestSpans } from 'opentelemetry-instrumentation-testing-utils';
 
 const instrumentation = new KafkaJsInstrumentation();
 
@@ -26,13 +25,7 @@ import {
 import { DummyPropagation } from './DummyPropagation';
 
 describe('instrumentation-kafkajs', () => {
-    const provider = new NodeTracerProvider();
-    const memoryExporter = new InMemorySpanExporter();
-    const spanProcessor = new SimpleSpanProcessor(memoryExporter);
     propagation.setGlobalPropagator(new DummyPropagation());
-    provider.addSpanProcessor(spanProcessor);
-    instrumentation.setTracerProvider(provider);
-    let contextManager: ContextManager;
 
     const kafka = new Kafka({
         clientId: 'unit-tests',
@@ -78,13 +71,6 @@ describe('instrumentation-kafkajs', () => {
 
     beforeEach(() => {
         messagesSent = [];
-        contextManager = new AsyncHooksContextManager().enable();
-        context.setGlobalContextManager(contextManager);
-    });
-
-    afterEach(() => {
-        memoryExporter.reset();
-        contextManager.disable();
     });
 
     describe('producer', () => {
@@ -95,19 +81,17 @@ describe('instrumentation-kafkajs', () => {
 
         describe('successful send', () => {
             beforeEach(async () => {
-                patchProducerSend(
-                    async (): Promise<RecordMetadata[]> => {
-                        return [
-                            {
-                                topicName: 'topic-name-1',
-                                partition: 0,
-                                errorCode: 123,
-                                offset: '18',
-                                timestamp: '123456',
-                            },
-                        ];
-                    }
-                );
+                patchProducerSend(async (): Promise<RecordMetadata[]> => {
+                    return [
+                        {
+                            topicName: 'topic-name-1',
+                            partition: 0,
+                            errorCode: 123,
+                            offset: '18',
+                            timestamp: '123456',
+                        },
+                    ];
+                });
                 instrumentation.disable();
                 instrumentation.enable();
                 producer = kafka.producer();
@@ -126,7 +110,7 @@ describe('instrumentation-kafkajs', () => {
                 expect(res.length).toBe(1);
                 expect(res[0].topicName).toStrictEqual('topic-name-1');
 
-                const spans = memoryExporter.getFinishedSpans();
+                const spans = getTestSpans();
                 expect(spans.length).toBe(1);
                 const span = spans[0];
                 expect(span.kind).toStrictEqual(SpanKind.PRODUCER);
@@ -155,7 +139,7 @@ describe('instrumentation-kafkajs', () => {
                     ],
                 });
 
-                const spans = memoryExporter.getFinishedSpans();
+                const spans = getTestSpans();
                 expect(spans.length).toBe(2);
                 expect(spans[0].name).toStrictEqual('topic-name-1');
                 expect(spans[1].name).toStrictEqual('topic-name-1');
@@ -190,7 +174,7 @@ describe('instrumentation-kafkajs', () => {
                     ],
                 });
 
-                const spans = memoryExporter.getFinishedSpans();
+                const spans = getTestSpans();
                 expect(spans.length).toBe(3);
                 expect(spans[0].name).toStrictEqual('topic-name-1');
                 expect(spans[1].name).toStrictEqual('topic-name-1');
@@ -205,11 +189,9 @@ describe('instrumentation-kafkajs', () => {
 
         describe('failed send', () => {
             beforeEach(async () => {
-                patchProducerSend(
-                    (): Promise<RecordMetadata[]> => {
-                        return Promise.reject(new Error('error thrown from kafka client send'));
-                    }
-                );
+                patchProducerSend((): Promise<RecordMetadata[]> => {
+                    return Promise.reject(new Error('error thrown from kafka client send'));
+                });
                 instrumentation.disable();
                 instrumentation.enable();
                 producer = kafka.producer();
@@ -227,7 +209,7 @@ describe('instrumentation-kafkajs', () => {
                     });
                 } catch (err) {}
 
-                const spans = memoryExporter.getFinishedSpans();
+                const spans = getTestSpans();
                 expect(spans.length).toBe(1);
                 const span = spans[0];
                 expect(span.status.code).toStrictEqual(SpanStatusCode.ERROR);
@@ -249,7 +231,7 @@ describe('instrumentation-kafkajs', () => {
                     });
                 } catch (err) {}
 
-                const spans = memoryExporter.getFinishedSpans();
+                const spans = getTestSpans();
                 expect(spans.length).toBe(2);
                 spans.forEach((span) => {
                     expect(span.status.code).toStrictEqual(SpanStatusCode.ERROR);
@@ -284,7 +266,7 @@ describe('instrumentation-kafkajs', () => {
                     });
                 } catch (err) {}
 
-                const spans = memoryExporter.getFinishedSpans();
+                const spans = getTestSpans();
                 expect(spans.length).toBe(3);
                 spans.forEach((span) => {
                     expect(span.status.code).toStrictEqual(SpanStatusCode.ERROR);
@@ -318,7 +300,7 @@ describe('instrumentation-kafkajs', () => {
                     ],
                 });
 
-                const spans = memoryExporter.getFinishedSpans();
+                const spans = getTestSpans();
                 expect(spans.length).toBe(1);
                 const span = spans[0];
                 expect(span.attributes['attribute-from-hook']).toStrictEqual('testing message content');
@@ -350,7 +332,7 @@ describe('instrumentation-kafkajs', () => {
                     ],
                 });
 
-                const spans = memoryExporter.getFinishedSpans();
+                const spans = getTestSpans();
                 expect(spans.length).toBe(1);
                 const span = spans[0];
                 expect(span.status.code).toStrictEqual(SpanStatusCode.UNSET);
@@ -374,7 +356,7 @@ describe('instrumentation-kafkajs', () => {
                     messages: [{ value: 'testing message content' }],
                 });
 
-                const spans = memoryExporter.getFinishedSpans();
+                const spans = getTestSpans();
                 expect(spans.length).toBe(1);
                 const span = spans[0];
                 expect(span.attributes['module.version']).toMatch(/\d{1,4}\.\d{1,4}\.\d{1,5}.*/);
@@ -433,7 +415,7 @@ describe('instrumentation-kafkajs', () => {
                 const payload: EachMessagePayload = createEachMessagePayload();
                 await runConfig.eachMessage(payload);
 
-                const spans = memoryExporter.getFinishedSpans();
+                const spans = getTestSpans();
                 expect(spans.length).toBe(1);
                 const span = spans[0];
                 expect(span.name).toStrictEqual('topic-name-1');
@@ -459,7 +441,7 @@ describe('instrumentation-kafkajs', () => {
                 const payload: EachMessagePayload = createEachMessagePayload();
                 await runConfig.eachMessage(payload);
 
-                const spans = memoryExporter.getFinishedSpans();
+                const spans = getTestSpans();
                 expect(spans.length).toBe(1);
             });
         });
@@ -486,7 +468,7 @@ describe('instrumentation-kafkajs', () => {
                 const payload: EachMessagePayload = createEachMessagePayload();
                 await runConfig.eachMessage(payload);
 
-                const spans = memoryExporter.getFinishedSpans();
+                const spans = getTestSpans();
                 expect(spans.length).toBe(1);
                 const span = spans[0];
                 expect(span.attributes['attribute key from hook']).toStrictEqual(payload.message.value.toString());
@@ -515,7 +497,7 @@ describe('instrumentation-kafkajs', () => {
                 const payload: EachMessagePayload = createEachMessagePayload();
                 await runConfig.eachMessage(payload);
 
-                const spans = memoryExporter.getFinishedSpans();
+                const spans = getTestSpans();
                 // span should still be created
                 expect(spans.length).toBe(1);
             });
@@ -547,7 +529,7 @@ describe('instrumentation-kafkajs', () => {
                 }
                 expect(exception).toEqual(errorToThrow);
 
-                const spans = memoryExporter.getFinishedSpans();
+                const spans = getTestSpans();
                 expect(spans.length).toBe(1);
                 const span = spans[0];
                 expect(span.status.code).toStrictEqual(SpanStatusCode.ERROR);
@@ -573,7 +555,7 @@ describe('instrumentation-kafkajs', () => {
                 }
                 expect(exception).toEqual(objectToThrow);
 
-                const spans = memoryExporter.getFinishedSpans();
+                const spans = getTestSpans();
                 expect(spans.length).toBe(1);
                 const span = spans[0];
                 expect(span.status.code).toStrictEqual(SpanStatusCode.ERROR);
@@ -596,7 +578,7 @@ describe('instrumentation-kafkajs', () => {
                 }
                 expect(exception).toBeUndefined();
 
-                const spans = memoryExporter.getFinishedSpans();
+                const spans = getTestSpans();
                 expect(spans.length).toBe(1);
                 const span = spans[0];
                 expect(span.status.code).toStrictEqual(SpanStatusCode.ERROR);
@@ -620,7 +602,7 @@ describe('instrumentation-kafkajs', () => {
                 const payload: EachBatchPayload = createEachBatchPayload();
                 await runConfig.eachBatch(payload);
 
-                const spans = memoryExporter.getFinishedSpans();
+                const spans = getTestSpans();
                 expect(spans.length).toBe(3);
                 spans.forEach((span) => {
                     expect(span.name).toStrictEqual('topic-name-1');
@@ -657,7 +639,7 @@ describe('instrumentation-kafkajs', () => {
                 const payload: EachBatchPayload = createEachBatchPayload();
                 await runConfig.eachBatch(payload);
 
-                const spans = memoryExporter.getFinishedSpans();
+                const spans = getTestSpans();
                 expect(spans.length).toBe(3);
             });
         });
@@ -682,7 +664,7 @@ describe('instrumentation-kafkajs', () => {
                 const payload: EachMessagePayload = createEachMessagePayload();
                 await runConfig.eachMessage(payload);
 
-                const spans = memoryExporter.getFinishedSpans();
+                const spans = getTestSpans();
                 expect(spans.length).toBe(1);
                 const span = spans[0];
                 expect(span.attributes['module.version']).toMatch(/\d{1,4}\.\d{1,4}\.\d{1,5}.*/);
@@ -730,7 +712,7 @@ describe('instrumentation-kafkajs', () => {
             };
             await runConfig.eachMessage(consumerPayload);
 
-            const spans = memoryExporter.getFinishedSpans();
+            const spans = getTestSpans();
             expect(spans.length).toBe(2);
             const [producerSpan, consumerSpan] = spans;
             expect(consumerSpan.spanContext.traceId).toStrictEqual(producerSpan.spanContext.traceId);
@@ -772,7 +754,7 @@ describe('instrumentation-kafkajs', () => {
             } as EachBatchPayload;
             await runConfig.eachBatch(consumerPayload);
 
-            const spans = memoryExporter.getFinishedSpans();
+            const spans = getTestSpans();
             expect(spans.length).toBe(3);
             const [producerSpan, receivingSpan, processingSpan] = spans;