+++ packages/instrumentation-express/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.5](https://github.com/aspecto-io/opentelemetry-ext-js/compare/opentelemetry-instrumentation-express@0.4.4...opentelemetry-instrumentation-express@0.4.5) (2021-05-31)
+
+**Note:** Version bump only for package opentelemetry-instrumentation-express
+
+
+
+
+
+## [0.4.4](https://github.com/aspecto-io/opentelemetry-ext-js/compare/opentelemetry-instrumentation-express@0.4.3...opentelemetry-instrumentation-express@0.4.4) (2021-05-19)
+
+**Note:** Version bump only for package opentelemetry-instrumentation-express
+
+
+
+
+
+## [0.4.3](https://github.com/aspecto-io/aspecto-opentelemetry-js/compare/opentelemetry-instrumentation-express@0.4.2...opentelemetry-instrumentation-express@0.4.3) (2021-05-19)
+
+**Note:** Version bump only for package opentelemetry-instrumentation-express
+
+
+
+
+
 ## [0.4.2](https://github.com/aspecto-io/aspecto-opentelemetry-js/compare/opentelemetry-instrumentation-express@0.4.1...opentelemetry-instrumentation-express@0.4.2) (2021-05-04)
 
 **Note:** Version bump only for package opentelemetry-instrumentation-express