+++ packages/instrumentation-typeorm/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.3](https://github.com/aspecto-io/opentelemetry-ext-js/compare/opentelemetry-instrumentation-typeorm@0.4.2...opentelemetry-instrumentation-typeorm@0.4.3) (2021-05-31)
+
+**Note:** Version bump only for package opentelemetry-instrumentation-typeorm
+
+
+
+
+
+## [0.4.2](https://github.com/aspecto-io/opentelemetry-ext-js/compare/opentelemetry-instrumentation-typeorm@0.4.1...opentelemetry-instrumentation-typeorm@0.4.2) (2021-05-19)
+
+**Note:** Version bump only for package opentelemetry-instrumentation-typeorm
+
+
+
+
+
+## [0.4.1](https://github.com/aspecto-io/opentelemetry-ext-js/compare/opentelemetry-instrumentation-typeorm@0.4.0...opentelemetry-instrumentation-typeorm@0.4.1) (2021-05-19)
+
+**Note:** Version bump only for package opentelemetry-instrumentation-typeorm
+
+
+
+
+
 # [0.4.0](https://github.com/aspecto-io/opentelemetry-ext-js/compare/opentelemetry-instrumentation-typeorm@0.3.0...opentelemetry-instrumentation-typeorm@0.4.0) (2021-04-28)
 
 **Note:** Version bump only for package opentelemetry-instrumentation-typeorm